+++ apps/backend/src/modules/chat-messages/libs/enums/relation-name.enum.ts
@@ -1,0 +1,6 @@
+const RelationName = {
+	SENDER_USER: "senderUser",
+	USER_DETAILS: "userDetails",
+} as const;
+
+export { RelationName };