+++ apps/backend/src/modules/chats/libs/types/types.ts
@@ -1,0 +1,6 @@
+export {
+	type ChatCreateRequestDto,
+	type ChatGetAllItemResponseDto,
+	type ChatItemResponseDto,
+	type ChatResponseDto,
+} from "@trackmates/shared";