+++ apps/frontend/src/index.tsx
@@ -18,6 +18,7 @@ import { Friends } from "./pages/friends/friends.js";
 import { NotFound } from "./pages/not-found/not-found.js";
 import { Overview } from "./pages/overview/overview.js";
 import { Profile } from "./pages/profile/profile.js";
+import { User } from "./pages/user/user.js";
 
 createRoot(document.querySelector("#root") as HTMLElement).render(
 	<StrictMode>
@@ -58,6 +59,10 @@ createRoot(document.querySelector("#root") as HTMLElement).render(
 							{
 								element: <ProtectedRoute component={<CourseInfo />} />,
 								path: AppRoute.COURSE_INFO_$ID,
+							},
+							{
+								element: <ProtectedRoute component={<User />} />,
+								path: AppRoute.USERS_$ID,
 							},
 						],
 						element: <App />,