+++ apps/frontend/src/modules/courses/slices/courses.slice.ts
@@ -4,16 +4,18 @@ import { DataStatus } from "~/libs/enums/enums.js";
 import { type ValueOf } from "~/libs/types/types.js";
 
 import { type CourseDto } from "../libs/types/types.js";
-import { getAll, getById } from "./actions.js";
+import { getAll, getById, getRecommended } from "./actions.js";
 
 type State = {
 	currentCourse: CourseDto | null;
+	recommendedCourses: CourseDto[];
 	searchDataStatus: ValueOf<typeof DataStatus>;
 	searchedCourses: CourseDto[];
 };
 
 const initialState: State = {
 	currentCourse: null,
+	recommendedCourses: [],
 	searchDataStatus: DataStatus.IDLE,
 	searchedCourses: [],
 };
@@ -38,6 +40,16 @@ const { actions, name, reducer } = createSlice({
 			state.searchDataStatus = DataStatus.PENDING;
 		});
 		builder.addCase(getById.rejected, (state) => {
+			state.searchDataStatus = DataStatus.REJECTED;
+		});
+		builder.addCase(getRecommended.fulfilled, (state, action) => {
+			state.recommendedCourses = action.payload.courses;
+			state.searchDataStatus = DataStatus.FULFILLED;
+		});
+		builder.addCase(getRecommended.pending, (state) => {
+			state.searchDataStatus = DataStatus.PENDING;
+		});
+		builder.addCase(getRecommended.rejected, (state) => {
 			state.searchDataStatus = DataStatus.REJECTED;
 		});
 	},