+++ apps/frontend/src/modules/courses/slices/courses.ts
@@ -1,10 +1,11 @@
-import { getAll, getById } from "./actions.js";
+import { getAll, getById, getRecommended } from "./actions.js";
 import { actions } from "./courses.slice.js";
 
 const allActions = {
 	...actions,
 	getAll,
 	getById,
+	getRecommended,
 };
 
 export { allActions as actions };