+++ apps/frontend/src/modules/users/slices/users.slice.ts
@@ -1,9 +1,37 @@
 import { createSlice } from "@reduxjs/toolkit";
 
-const { name } = createSlice({
-	initialState: {},
+import { DataStatus } from "~/libs/enums/enums.js";
+import { type ValueOf } from "~/libs/types/types.js";
+
+import { type UserAuthResponseDto } from "../libs/types/types.js";
+import { getById } from "./actions.js";
+
+type State = {
+	dataStatus: ValueOf<typeof DataStatus>;
+	profileUser: UserAuthResponseDto | null;
+};
+
+const initialState: State = {
+	dataStatus: DataStatus.IDLE,
+	profileUser: null,
+};
+
+const { actions, name, reducer } = createSlice({
+	extraReducers(builder) {
+		builder.addCase(getById.fulfilled, (state, action) => {
+			state.profileUser = action.payload;
+			state.dataStatus = DataStatus.FULFILLED;
+		});
+		builder.addCase(getById.pending, (state) => {
+			state.dataStatus = DataStatus.PENDING;
+		});
+		builder.addCase(getById.rejected, (state) => {
+			state.dataStatus = DataStatus.REJECTED;
+		});
+	},
+	initialState,
 	name: "users",
 	reducers: {},
 });
 
-export { name };
+export { actions, name, reducer };