+++ apps/frontend/src/pages/course-info/components/course-sections/libs/constants.ts
@@ -1,0 +1,4 @@
+const MAXIMUM_ITEMS = 10;
+const START_INDEX = 0;
+
+export { MAXIMUM_ITEMS, START_INDEX };