+++ apps/frontend/src/pages/friends/libs/components/friend/friend.tsx
@@ -1,11 +1,14 @@
 import friendImage from "~/assets/img/friend.jpeg";
-import { Button, Image } from "~/libs/components/components.js";
+import { Button, Image, Link } from "~/libs/components/components.js";
+import { type AppRoute } from "~/libs/enums/enums.js";
 import {
 	useAppDispatch,
 	useAppSelector,
 	useCallback,
+	useEffect,
 	useState,
 } from "~/libs/hooks/hooks.js";
+import { type ValueOf } from "~/libs/types/types.js";
 import { actions } from "~/modules/friends/friends.js";
 import { type UserAuthResponseDto } from "~/modules/users/users.js";
 
@@ -21,7 +24,15 @@ const Friend: React.FC<Properties> = ({ friend }: Properties) => {
 			state.friends.followings.some((user) => user.id === friend.id),
 		),
 	);
+
+	const isFollowingFromSelector = useAppSelector((state) =>
+		state.friends.followings.some((user) => user.id === friend.id),
+	);
 	const dispatch = useAppDispatch();
+
+	useEffect(() => {
+		setIsFollowing(isFollowingFromSelector);
+	}, [isFollowingFromSelector]);
 
 	const handleFollow = useCallback(() => {
 		void dispatch(actions.follow({ id: friend.id }))
@@ -41,16 +52,19 @@ const Friend: React.FC<Properties> = ({ friend }: Properties) => {
 
 	return (
 		<article className={styles["card"]}>
-			<div className={styles["card-content"]}>
+			<Link
+				className={styles["card-content"]}
+				to={`/users/${friend.id}` as ValueOf<typeof AppRoute>}
+			>
 				<Image
 					alt="User avatar"
 					className={styles["portrait"]}
 					src={friendImage}
 				/>
-				<p className={styles["fullName"]}>
-					{friend.firstName} {friend.lastName}
-				</p>
-			</div>
+				<p
+					className={styles["fullName"]}
+				>{`${friend.firstName} ${friend.lastName}`}</p>
+			</Link>
 
 			<div className={styles["actions"]}>
 				<Button