+++ packages/shared/src/modules/chat-messages/libs/types/chat-message-create-request-dto.type.ts
@@ -1,0 +1,6 @@
+type ChatMessageCreateRequestDto = {
+	chatId: number;
+	text: string;
+};
+
+export { type ChatMessageCreateRequestDto };