+++ packages/shared/src/modules/chat-messages/libs/types/chat-message-update-request-dto.type.ts
@@ -1,0 +1,5 @@
+type ChatMessageUpdateRequestDto = {
+	text: string;
+};
+
+export { type ChatMessageUpdateRequestDto };